+++ backend/internal/domain/repo/jobs/product.go
@@ -7,9 +7,9 @@ import (
 )
 
 type ProductRepository interface {
-	First(ctx context.Context, id int) (*jobs.JobProduct, error)
-	Create(ctx context.Context, jobProduct *jobs.JobProduct) (int, error)
-	UpdateJobTime(ctx context.Context, id int) error
-	UpdateStatus(ctx context.Context, id int, status int) error
+	First(ctx context.Context, id int64) (*jobs.JobProduct, error)
+	Create(ctx context.Context, jobProduct *jobs.JobProduct) (int64, error)
+	UpdateJobTime(ctx context.Context, id int64) error
+	UpdateStatus(ctx context.Context, id int64, status int) error
 	Clear(ctx context.Context) error
 }
